Fix enterprise upgrade tests
