+++ .github/workflows/e2e.yml
@@ -101,8 +101,6 @@ jobs:
           distro: ubi
         - test-name: replica_set_recovery
           distro: ubi
-        - test-name: replica_set_recovery
-          distro: ubi
         - test-name: replica_set_mongod_readiness
           distro: ubi
         - test-name: replica_set_scale
@@ -116,8 +114,6 @@ jobs:
         - test-name: prometheus
           distro: ubi
         - test-name: replica_set_tls
-          distro: ubi
-        - test-name: replica_set_tls_pem_file
           distro: ubi
         - test-name: replica_set_tls_recreate_mdbc
           distro: ubi
