+++ test/e2e/replica_set_enterprise_upgrade/replica_set_enterprise_upgrade.go
@@ -1,38 +1,20 @@
-package replica_set
+package replica_set_enterprise_upgrade
 
 import (
 	"fmt"
-	"os"
 	"testing"
 	"time"
 
 	"github.com/mongodb/mongodb-kubernetes-operator/controllers/construct"
-
-	. "github.com/mongodb/mongodb-kubernetes-operator/test/e2e/util/mongotester"
-
-	e2eutil "github.com/mongodb/mongodb-kubernetes-operator/test/e2e"
+	"github.com/mongodb/mongodb-kubernetes-operator/test/e2e"
 	"github.com/mongodb/mongodb-kubernetes-operator/test/e2e/mongodbtests"
-	setup "github.com/mongodb/mongodb-kubernetes-operator/test/e2e/setup"
+	"github.com/mongodb/mongodb-kubernetes-operator/test/e2e/setup"
+	"github.com/mongodb/mongodb-kubernetes-operator/test/e2e/util/mongotester"
 )
-
-var (
-	versionsForUpgrades = []string{"4.4.19", "5.0.15"}
-)
-
-func TestMain(m *testing.M) {
-	code, err := e2eutil.RunTest(m)
-	if err != nil {
-		fmt.Println(err)
-	}
-	os.Exit(code)
-}
-
-func TestReplicaSet(t *testing.T) {
-	DeployEnterpriseAndUpgradeTest(t, versionsForUpgrades)
-}
 
 func DeployEnterpriseAndUpgradeTest(t *testing.T, versionsToBeTested []string) {
-	t.Setenv(construct.MongodbName, "mongodb-enterprise-server")
+	t.Setenv(construct.MongodbRepoUrl, "docker.io/mongodb")
+	t.Setenv(construct.MongodbImageEnv, "mongodb-enterprise-server")
 	ctx := setup.Setup(t)
 	defer ctx.Teardown()
 
@@ -44,7 +26,7 @@ func DeployEnterpriseAndUpgradeTest(t *testing.T, versionsToBeTested []string) {
 		t.Fatal(err)
 	}
 
-	tester, err := FromResource(t, mdb)
+	tester, err := mongotester.FromResource(t, mdb)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -55,9 +37,9 @@ func DeployEnterpriseAndUpgradeTest(t *testing.T, versionsToBeTested []string) {
 	t.Run("AutomationConfig has the correct version", mongodbtests.AutomationConfigVersionHasTheExpectedVersion(&mdb, 1))
 
 	for i := 1; i < len(versionsToBeTested); i++ {
-		t.Run(fmt.Sprintf("Testing upgrade from %s to %s", versionsForUpgrades[i-1], versionsForUpgrades[i]), func(t *testing.T) {
+		t.Run(fmt.Sprintf("Testing upgrade from %s to %s", versionsToBeTested[i-1], versionsToBeTested[i]), func(t *testing.T) {
 			defer tester.StartBackgroundConnectivityTest(t, time.Second*10)()
-			t.Run(fmt.Sprintf("Upgrading to %s", versionsForUpgrades[i]), mongodbtests.ChangeVersion(&mdb, versionsForUpgrades[i]))
+			t.Run(fmt.Sprintf("Upgrading to %s", versionsToBeTested[i]), mongodbtests.ChangeVersion(&mdb, versionsToBeTested[i]))
 			t.Run("Stateful Set Reaches Ready State, after Upgrading", mongodbtests.StatefulSetBecomesReady(&mdb))
 			t.Run("MongoDB Reaches Running Phase", mongodbtests.MongoDBReachesRunningPhase(&mdb))
 			t.Run("AutomationConfig's version has been increased", mongodbtests.AutomationConfigVersionHasTheExpectedVersion(&mdb, i+1))
