+++ test/e2e/replica_set_enterprise_upgrade_4_5/replica_set_enterprise_upgrade_4_5_test.go
@@ -1,0 +1,27 @@
+package replica_set
+
+import (
+	"fmt"
+	"os"
+	"testing"
+
+	"github.com/mongodb/mongodb-kubernetes-operator/test/e2e/replica_set_enterprise_upgrade"
+
+	e2eutil "github.com/mongodb/mongodb-kubernetes-operator/test/e2e"
+)
+
+var (
+	versionsForUpgrades = []string{"4.4.19", "5.0.15"}
+)
+
+func TestMain(m *testing.M) {
+	code, err := e2eutil.RunTest(m)
+	if err != nil {
+		fmt.Println(err)
+	}
+	os.Exit(code)
+}
+
+func TestReplicaSet(t *testing.T) {
+	replica_set_enterprise_upgrade.DeployEnterpriseAndUpgradeTest(t, versionsForUpgrades)
+}
